+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Ics.py changelog
  - Adds support for Organizers to Events
  - Adds support for Attendees to Events
  - Added LAST-MODIFIED attribute support
+ - Event equality now checks all fields (except uid)
 
 **************
 0.4
@@ -40,7 +41,7 @@ Misc:
     - Multi-day events fixes (thanks @ConnyOnny)
     - Fix `TZID` drop when `VTIMEZONE` is empty (thanks @ConnyOnny)
     - Better test coverage (thanks @aureooms)
-    
+
 Breaking Changes:
     - Removed EventList class